backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
pandas
scikit-learn
joblib
requests
google-generativeai
python-dotenv
google-generativeai
